% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2neo4j.R
\name{delete_all_neo4j}
\alias{delete_all_neo4j}
\title{delete_all_neo4j}
\usage{
delete_all_neo4j()
}
\description{
WARNING: This function deletes all nodes and relationships in your current Neo4j database.
}
\details{
WARNING: This function deletes all nodes and relationships in your current Neo4j database.
This comes in handy if you're developing a repeatable process to build your graph database and you want
to repeatedly experiment and return to a blank slate.
}
