% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2neo4j.R
\name{enhance_from_keys}
\alias{enhance_from_keys}
\title{enhance_from_keys}
\usage{
enhance_from_keys(
  target_df,
  source_key_col,
  source_prop_col,
  target_key_col_name,
  target_prop_name,
  remove_target_key_col = TRUE
)
}
\arguments{
\item{target_df}{The data.frame to be enhanced.}

\item{source_key_col}{The column in the reference data.frame containing
the keys/IDs, passed as a vector, eg. Employees$id}

\item{source_prop_col}{The column in the reference data.frame containing
the values, passed as a vector, eg. Employees$name}

\item{target_key_col_name}{The name of the column in the target data.frame
containing the keys/IDs, passed as a string, eg. "Employee_ID"}

\item{target_prop_name}{The name for the new column to be added to the
target data.frame containing the values, passed as a string,
eg. "Employee_Name"}
}
\description{
Enhance a target data.frame that has one column of keys or IDs from
the key-value pair that exists in two columns of a source data.frame.
}
\details{
This handles a join to enhance one data.frame that contains a column
of keys/IDs from another data.frame that contains columns of key-value
pairs. For example, to lookup names by reference to ID numbers.
}
