% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2neo4j.R
\name{neo.find.nodes}
\alias{neo.find.nodes}
\title{neo.find.nodes}
\usage{
neo.find.nodes(
  search_term,
  var = "n",
  label = "",
  propery = "any_property",
  exact = FALSE,
  return = ""
)
}
\arguments{
\item{search_term}{Term to search for in node properties. Not case-sensitive.}

\item{var}{Variable to assign to this node. Defaults to "n"}

\item{label}{To search only nodes with a particular label, eg "Person". Defaults to any label.}

\item{exact}{Use TRUE if you want only exact matches of your search term. Defaults to FALSE to match any property value that CONTAINS your search term.}

\item{return}{Used to construct the RETURN portion of the Cypher query.}

\item{property}{The property type to search in. Defaults to any property.}
}
\value{
Cypher query (or partial query) as a string to pass to Neo4j
}
\description{
Helps to construct Cypher statements to search for matching nodes in Neo4j.
}
\details{

}
\examples{

bn <- neo.find.nodes("Brendan")

}
