% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2neo4j.R
\name{neo.rel}
\alias{neo.rel}
\title{neo.rel}
\usage{
neo.rel(rel_var = "r", relationship = "", degrees = "1", directed = "--")
}
\arguments{
\item{rel_var}{To assign a variable to the relationship. Defaults to "r".}

\item{relationship}{The relationship label, eg. "Knows".}

\item{degrees}{Degrees of separation defaults to 1. For a range use eg. "1..4" for nodes connected within 1 to 4 degrees of distance.}

\item{directed}{Directed defaults to undirected, but can be changed to "r", "-->" or "right" for right-directed, eg. a --> b (or similar for left-directed).}
}
\value{
A string describing a relationship in Cypher
}
\description{
Helps to construct relationship statements for people not familiar with Cypher syntax
}
\details{

}
\examples{

}
