% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2neo4j.R
\name{neo.search}
\alias{neo.search}
\title{neo.search}
\usage{
neo.search(
  a.search_term,
  a.var = "a",
  a.label,
  a.property,
  a.exact,
  b.search_term,
  b.var = "b",
  b.label,
  b.property,
  b.exact,
  rel.var = "r",
  rel.relationship,
  rel.degrees,
  rel.directed
)
}
\arguments{
\item{a.search_term}{Term to search for in node properties. Not case-sensitive.}

\item{a.var}{Variable to assign to this node. Defaults to "n"}

\item{a.label}{To search only nodes with a particular label, eg "Person". Defaults to any label.}

\item{a.property}{The property type to search in. Defaults to any property.}

\item{a.exact}{Use TRUE if you want only exact matches of your search term. Defaults to FALSE to match any property value that CONTAINS your search term.}

\item{rel_var}{To assign a variable to the relationship. Defaults to "r".}

\item{relationship}{The relationship label.}

\item{degrees}{Degrees of separation defaults to 1. FOr a range use eg. "1..4"}

\item{directed}{Directed defaults to undirected, but can be changed to "r", "-->" or "right" for right-directed, eg. a --> b (or similar for left-directed).}
}
\value{
A relationship defined as a Cypher string.
}
\description{
Get results for a query involving two nodes and the relationship(s) or path between them
}
\details{
The arguments below are the same for the first node, a, and the second node, b. (Just change the letter prefix.)
}
\examples{

}
