% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2neo4j.R
\name{nodify_cols}
\alias{nodify_cols}
\title{nodify_cols}
\usage{
nodify_cols(df, col, label = "df_col_name", property_key = "Name", sep = ";")
}
\arguments{
\item{df}{The data.frame containing the column of data.}

\item{col}{The names of the data columns, as a list or vector, eg. col = c("Organization", "City", "Postal_Code")}

\item{label}{The label to assign to these nodes. The default is to take this
label from the name of the column. As a string.}

\item{property_key}{The name for the property key for these values. Eg. "Name" for nodes labeled "Person". The default is "Name". As a string.}

\item{sep}{The separator between items in a cell of the column, eg "," or ";"}
}
\description{
A shortcut to run nodify_col on multiple columns. The only difference is to
enter the column names as a character vector.
}
\details{
Assuming all defaults, with lots of simple, unique-name-only columns representing
nodes that have the node label as column name, and will have unique prop called Name (or
whatever you set for all of them)

ie. in Cypher: (:Label) and Label.Name
}
